Fix ajax POST for a poem
This fixes a bug where the `id` field of a poem was being converted to a
string after being sent to the server and before being saved to a file.
This behavior is fixed when both re-ordering poems and when adding a new
poem. Also, the server now uses `bodyParser.json` and so expects request
bodies to be json.